Add files via upload
To get everything to work properly, put the gun on the Interactable layer, put the player on the Player layer, and set the Bullet under the prefabs folder to the Bullet layer. Then go to Edit->Project Settings->Physics2D uncheck the collision layer that corresponds to Bullet and Player to make it so the bullet doesn't interact with the player.